+++ upcloud/service/utils_test.go
@@ -2,7 +2,9 @@ package service
 
 import (
 	"context"
+	"encoding/csv"
 	"encoding/json"
+	"io"
 	"log"
 	"net/http"
 	"net/http/httptest"
@@ -93,17 +95,73 @@ func record(t *testing.T, fixture string, f func(context.Context, *testing.T, *r
 		// Redact sensitive information from response body
 		if i.Response.Body != "" {
 			var responseData map[string]interface{}
+			var responseArrayData []map[string]any
+
+			const maxTopLevelRecords = 100
+			const redactedToken = "ucat_[REDACTED]" //#nosec G101 -- false positive
 
 			err := json.Unmarshal([]byte(i.Response.Body), &responseData)
 			if err == nil {
 				// Redact sensitive fields
 				if _, exists := responseData["token"]; exists {
-					responseData["token"] = "ucat_[REDACTED]"
+					responseData["token"] = redactedToken
 				}
 
 				// Convert back to string and update response body
-				updatedBody, _ := json.Marshal(responseData)
-				i.Response.Body = string(updatedBody)
+				if updatedBody, err := json.Marshal(responseData); err != nil {
+					return err
+				} else {
+					i.Response.Body = string(updatedBody)
+				}
+			} else if err = json.Unmarshal([]byte(i.Response.Body), &responseArrayData); err == nil {
+				// Truncate some to reduce fixture size
+				if len(responseArrayData) > maxTopLevelRecords {
+					responseArrayData = responseArrayData[:maxTopLevelRecords]
+				}
+
+				// Redact sensitive fields
+				for _, responseData = range responseArrayData {
+					if _, exists := responseData["token"]; exists {
+						responseData["token"] = redactedToken
+					}
+				}
+
+				// Convert back to string and update response body
+				if updatedBody, err := json.Marshal(responseArrayData); err != nil {
+					return err
+				} else {
+					i.Response.Body = string(updatedBody)
+				}
+			} else if strings.HasPrefix(i.Response.Headers.Get("Content-Type"), "text/csv") {
+				cr := csv.NewReader(strings.NewReader(i.Response.Body))
+				sb := &strings.Builder{}
+				cw := csv.NewWriter(sb)
+				n := 0
+				for n < maxTopLevelRecords { // Truncate some to reduce fixture size
+					record, err := cr.Read()
+					if err != nil {
+						return err
+					}
+					if err == io.EOF {
+						break
+					}
+					n++
+
+					// Redact sensitive fields
+					// TODO: redaction based on column name
+					for j, field := range record {
+						if strings.HasPrefix(field, "ucat_") {
+							record[j] = redactedToken
+						}
+					}
+
+					if err = cw.Write(record); err != nil {
+						return err
+					}
+				}
+
+				// Convert back to string and update response body
+				i.Response.Body = sb.String()
 			}
 		}