+++ CHANGELOG.md
@@ -9,6 +9,15 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 - loadbalancer: `WaitForLoadBalancerOperationalState` helper for waiting for the load balancer to achieve a desired operational state
 - loadbalancer: `WaitForLoadBalancerDeletion` blocks execution until the specified load balancer instance has been deleted.
 
+## [8.20.0]
+
+### Added
+- audit logs: export support
+- client: streaming Get/Do variants
+
+### Changed
+- client: set minimum TLS version to 1.3
+
 ## [8.19.0]
 
 ### Added
@@ -593,7 +602,8 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 First stable release
 
-[Unreleased]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.19.0...HEAD
+[Unreleased]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.20.0...HEAD
+[8.20.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.19.0...v8.20.0
 [8.19.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.18.0...v8.19.0
 [8.18.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.17.0...v8.18.0
 [8.17.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.16.1...v8.17.0